+++ src/otx/algorithms/classification/adapters/mmcls/configurer.py
@@ -3,23 +3,20 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-import importlib
 from typing import Optional
 
 import torch
 from mmcv import build_from_cfg
-from mmcv.utils import Config, ConfigDict
+from mmcv.utils import ConfigDict
 
 from otx.algorithms import TRANSFORMER_BACKBONES
 from otx.algorithms.classification.adapters.mmcls.utils import (
     patch_datasets,
     patch_evaluation,
 )
+from otx.algorithms.common.adapters.mmcv.clsincr_mixin import IncrConfigurerMixin
 from otx.algorithms.common.adapters.mmcv.configurer import BaseConfigurer
-from otx.algorithms.common.adapters.mmcv.utils import (
-    build_dataloader,
-    build_dataset,
-)
+from otx.algorithms.common.adapters.mmcv.semisl_mixin import SemiSLConfigurerMixin
 from otx.algorithms.common.adapters.mmcv.utils.config_utils import (
     InputSizeManager,
     get_configured_input_size,
@@ -36,42 +33,19 @@ logger = get_logger()
 class ClassificationConfigurer(BaseConfigurer):
     """Patch config to support otx train."""
 
-    # pylint: disable=too-many-arguments
-    def configure(
-        self,
-        cfg,
-        model_ckpt,
-        data_cfg,
-        ir_options=None,
-        data_classes=None,
-        model_classes=None,
-        input_size: InputSizePreset = InputSizePreset.DEFAULT,
-        **kwargs,
-    ):
-        """Create MMCV-consumable config from given inputs."""
-        logger.info(f"configure!: training={self.training}")
-
-        self.configure_base(cfg, data_cfg, data_classes, model_classes, **kwargs)
-        self.configure_device(cfg)
-        self.configure_ckpt(cfg, model_ckpt)
-        self.configure_model(cfg, ir_options)
-        self.configure_data(cfg, data_cfg)
-        self.configure_input_size(cfg, input_size, model_ckpt)
-        self.configure_task(cfg)
-        self.configure_hook(cfg)
-        self.configure_samples_per_gpu(cfg)
-        self.configure_fp16(cfg)
-        self.configure_compat_cfg(cfg)
-        return cfg
-
-    def configure_compatibility(self, cfg, **kwargs):
-        """Configure for OTX compatibility with mmcls."""
+    def configure_data_pipeline(self, cfg, input_size, model_ckpt_path, **kwargs):
+        """Configuration for data pipeline."""
+        super().configure_data_pipeline(cfg, input_size, model_ckpt_path)
         options_for_patch_datasets = kwargs.get("options_for_patch_datasets", {"type": "OTXClsDataset"})
-        options_for_patch_evaluation = kwargs.get("options_for_patch_evaluation", {"task": "normal"})
         patch_datasets(cfg, **options_for_patch_datasets)
-        patch_evaluation(cfg, **options_for_patch_evaluation)
 
-    def configure_model(self, cfg, ir_options):  # noqa: C901
+    def configure_recipe(self, cfg, **kwargs):
+        """Configuration for training recipe."""
+        options_for_patch_evaluation = kwargs.get("options_for_patch_evaluation", {"task": "normal"})
+        patch_evaluation(cfg, **options_for_patch_evaluation)
+        super().configure_recipe(cfg)
+
+    def configure_backbone(self, cfg, ir_options):  # noqa: C901
         """Patch config's model.
 
         Change model type to super type
@@ -189,10 +163,10 @@ class ClassificationConfigurer(BaseConfigurer):
 
     @staticmethod
     def configure_input_size(
-        cfg, input_size_config: InputSizePreset = InputSizePreset.DEFAULT, model_ckpt: Optional[str] = None
+        cfg, input_size_config: InputSizePreset = InputSizePreset.DEFAULT, model_ckpt_path: Optional[str] = None
     ):
         """Change input size if necessary."""
-        input_size = get_configured_input_size(input_size_config, model_ckpt)
+        input_size = get_configured_input_size(input_size_config, model_ckpt_path)
         if input_size is None:
             return
 
@@ -200,78 +174,20 @@ class ClassificationConfigurer(BaseConfigurer):
         logger.info("Input size is changed to {}".format(input_size))
 
 
-CLASS_INC_DATASET = [
-    "OTXClsDataset",
-    "OTXMultilabelClsDataset",
-    "MPAHierarchicalClsDataset",
-    "ClsTVDataset",
-]
-WEIGHT_MIX_CLASSIFIER = ["CustomImageClassifier"]
-
-
-class IncrClassificationConfigurer(ClassificationConfigurer):
+class IncrClassificationConfigurer(IncrConfigurerMixin, ClassificationConfigurer):
     """Patch config to support incremental learning for classification."""
 
-    def configure_task(self, cfg):
+    def configure_task(self, cfg, **kwargs):
         """Patch config to support incremental learning."""
-        super().configure_task(cfg)
-        if "task_adapt" in cfg and self.task_adapt_type == "mpa":
+        super().configure_task(cfg, **kwargs)
+        if "task_adapt" in cfg and self.task_adapt_type == "default_task_adapt":
             self.configure_task_adapt_hook(cfg)
-
-    def configure_task_adapt_hook(self, cfg):
-        """Add TaskAdaptHook for sampler."""
-        train_data_cfg = self.get_data_cfg(cfg, "train")
-        if self.training:
-            if train_data_cfg.type not in CLASS_INC_DATASET:
-                logger.warning(f"Class Incremental Learning for {train_data_cfg.type} is not yet supported!")
-            if "new_classes" not in train_data_cfg:
-                logger.warning('"new_classes" should be defined for incremental learning w/ current model.')
-
-            if cfg.model.type in WEIGHT_MIX_CLASSIFIER:
-                cfg.model.task_adapt = ConfigDict(
-                    src_classes=self.org_model_classes,
-                    dst_classes=self.model_classes,
-                )
-            else:
-                logger.warning(f"Weight mixing for {cfg.model.type} is not yet supported!")
-
-            train_data_cfg.classes = self.model_classes
-
-        is_multiclass = not cfg.model.get("multilabel", False) and not cfg.model.get("hierarchical", False)
-        if is_multiclass:
-            efficient_mode = cfg["task_adapt"].get("efficient_mode", True)
-            sampler_type = "balanced"
-            self.configure_loss(cfg)
-        else:
-            efficient_mode = cfg["task_adapt"].get("efficient_mode", False)
-            sampler_type = "cls_incr"
-
-        if len(set(self.org_model_classes) & set(self.model_classes)) == 0 or set(self.org_model_classes) == set(
-            self.model_classes
-        ):
-            sampler_flag = False
-        else:
-            sampler_flag = True
-
-        # Update Task Adapt Hook
-        task_adapt_hook = ConfigDict(
-            type="TaskAdaptHook",
-            src_classes=self.org_model_classes,
-            dst_classes=self.model_classes,
-            model_type=cfg.model.type,
-            sampler_flag=sampler_flag,
-            sampler_type=sampler_type,
-            efficient_mode=efficient_mode,
-            use_adaptive_repeat=is_multiclass,
-            priority="NORMAL",
-        )
-        update_or_add_custom_hook(cfg, task_adapt_hook)
+            if self._is_multiclass(cfg):
+                self.configure_loss(cfg)
 
     def configure_loss(self, cfg):
         """Patch classification loss."""
-        if len(set(self.org_model_classes) & set(self.model_classes)) == 0 or set(self.org_model_classes) == set(
-            self.model_classes
-        ):
+        if not self.is_incremental():
             cfg.model.head.loss = dict(type="CrossEntropyLoss", loss_weight=1.0)
         else:
             cfg.model.head.loss = ConfigDict(
@@ -284,49 +200,25 @@ class IncrClassificationConfigurer(ClassificationConfigurer):
             )
             update_or_add_custom_hook(cfg, ib_loss_hook)
 
+    def get_sampler_type(self, cfg):
+        """Return sampler type."""
+        if self._is_multiclass(cfg):
+            sampler_type = "balanced"
+        else:
+            sampler_type = "cls_incr"
+        return sampler_type
 
-class SemiSLClassificationConfigurer(ClassificationConfigurer):
-    """Patch config to support semi supervised learning for classification."""
+    def use_adaptive_repeat(self, cfg) -> bool:
+        """Return whether using adaptive repeat.
 
-    def configure_hook(self, cfg):
-        """Update cfg.custom_hooks."""
-        super().configure_hook(cfg)
-        # Set unlabeled data hook
-        if self.training:
-            if cfg.data.get("unlabeled", False) and cfg.data.unlabeled.get("otx_dataset", False):
-                self.configure_unlabeled_dataloader(cfg)
+        Currently, only multi class classification supports adaptive repeat.
+        """
+        return self._is_multiclass(cfg)
 
     @staticmethod
-    def configure_unlabeled_dataloader(cfg: Config):
-        """Patch for unlabled dataloader."""
+    def _is_multiclass(cfg) -> bool:
+        return not cfg.model.get("multilabel", False) and not cfg.model.get("hierarchical", False)
 
-        model_task = {"classification": "mmcls", "detection": "mmdet", "segmentation": "mmseg"}
-        if "unlabeled" in cfg.data:
-            task_lib_module = importlib.import_module(f"{model_task[cfg.model_task]}.datasets")
-            dataset_builder = getattr(task_lib_module, "build_dataset")
-            dataloader_builder = getattr(task_lib_module, "build_dataloader")
 
-            dataset = build_dataset(cfg, "unlabeled", dataset_builder, consume=True)
-            unlabeled_dataloader = build_dataloader(
-                dataset,
-                cfg,
-                "unlabeled",
-                dataloader_builder,
-                distributed=cfg.distributed,
-                consume=True,
-            )
-
-            custom_hooks = cfg.get("custom_hooks", [])
-            updated = False
-            for custom_hook in custom_hooks:
-                if custom_hook["type"] == "ComposedDataLoadersHook":
-                    custom_hook["data_loaders"] = [*custom_hook["data_loaders"], unlabeled_dataloader]
-                    updated = True
-            if not updated:
-                custom_hooks.append(
-                    ConfigDict(
-                        type="ComposedDataLoadersHook",
-                        data_loaders=unlabeled_dataloader,
-                    )
-                )
-            cfg.custom_hooks = custom_hooks
+class SemiSLClassificationConfigurer(SemiSLConfigurerMixin, ClassificationConfigurer):
+    """Patch config to support semi supervised learning for classification."""