+++ src/otx/algorithms/detection/adapters/mmdet/configurer.py
@@ -3,21 +3,16 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-import importlib
 from typing import Optional
 
-from mmcv.utils import Config, ConfigDict
+from mmcv.utils import ConfigDict
 
+from otx.algorithms.common.adapters.mmcv.clsincr_mixin import IncrConfigurerMixin
 from otx.algorithms.common.adapters.mmcv.configurer import BaseConfigurer
-from otx.algorithms.common.adapters.mmcv.utils import (
-    build_dataloader,
-    build_dataset,
-)
+from otx.algorithms.common.adapters.mmcv.semisl_mixin import SemiSLConfigurerMixin
 from otx.algorithms.common.adapters.mmcv.utils.config_utils import (
     InputSizeManager,
     get_configured_input_size,
-    patch_color_conversion,
-    update_or_add_custom_hook,
 )
 from otx.algorithms.common.configs.configuration_enums import InputSizePreset
 from otx.algorithms.common.utils.logger import get_logger
@@ -33,38 +28,10 @@ logger = get_logger()
 class DetectionConfigurer(BaseConfigurer):
     """Patch config to support otx train."""
 
-    # pylint: disable=too-many-arguments
-    def configure(
-        self,
-        cfg,
-        train_dataset,
-        model_ckpt,
-        data_cfg,
-        ir_options=None,
-        data_classes=None,
-        model_classes=None,
-        input_size: InputSizePreset = InputSizePreset.DEFAULT,
-    ):
-        """Create MMCV-consumable config from given inputs."""
-        logger.info(f"configure!: training={self.training}")
-
-        self.configure_base(cfg, data_cfg, data_classes, model_classes)
-        self.configure_device(cfg)
-        self.configure_model(cfg, ir_options)
-        self.configure_ckpt(cfg, model_ckpt)
-        self.configure_data(cfg, data_cfg)
-        self.configure_input_size(cfg, input_size, model_ckpt)
+    def configure_model(self, cfg, data_classes, model_classes, ir_options, **kwargs):
+        """Configuration for model config."""
+        super().configure_model(cfg, data_classes, model_classes, ir_options, **kwargs)
         self.configure_regularization(cfg)
-        self.configure_task(cfg, train_dataset)
-        self.configure_hook(cfg)
-        self.configure_samples_per_gpu(cfg)
-        self.configure_fp16(cfg)
-        self.configure_compat_cfg(cfg)
-        return cfg
-
-    def configure_compatibility(self, cfg, **kwargs):
-        """Configure for OTX compatibility with mmdet."""
-        patch_color_conversion(cfg)
 
     def configure_regularization(self, cfg):  # noqa: C901
         """Patch regularization parameters."""
@@ -85,15 +52,19 @@ class DetectionConfigurer(BaseConfigurer):
                 if "weight_decay" in cfg.optimizer:
                     cfg.optimizer.weight_decay = 0.0
 
-    def configure_task(self, cfg, train_dataset):
+    def configure_task(self, cfg, **kwargs):
         """Patch config to support training algorithm."""
-        super().configure_task(cfg)
+
+        assert "train_dataset" in kwargs
+        train_dataset = kwargs["train_dataset"]
+
+        super().configure_task(cfg, **kwargs)
         if "task_adapt" in cfg:
             if self.data_classes != self.model_classes:
                 self.configure_task_data_pipeline(cfg)
-            if cfg["task_adapt"].get("use_mpa_anchor", False):
+            if cfg["task_adapt"].get("use_adaptive_anchor", False):
                 self.configure_anchor(cfg, train_dataset)
-            if self.task_adapt_type == "mpa":
+            if self.task_adapt_type == "default_task_adapt":
                 self.configure_bbox_head(cfg)
 
     def configure_classes(self, cfg):
@@ -139,7 +110,10 @@ class DetectionConfigurer(BaseConfigurer):
         class_adapt_cfg = dict(type="AdaptClassLabels", src_classes=self.data_classes, dst_classes=self.model_classes)
         pipeline_cfg = tr_data_cfg.pipeline
         for i, operation in enumerate(pipeline_cfg):
-            if operation["type"] == "LoadAnnotationFromOTXDataset":  # insert just after this operation
+            if operation["type"] in [
+                "LoadAnnotationFromOTXDataset",
+                "LoadResizeDataFromOTXDataset",
+            ]:  # insert just after this operation
                 op_next_ann = pipeline_cfg[i + 1] if i + 1 < len(pipeline_cfg) else {}
                 if op_next_ann.get("type", "") == class_adapt_cfg["type"]:
                     op_next_ann.update(class_adapt_cfg)
@@ -184,10 +158,10 @@ class DetectionConfigurer(BaseConfigurer):
 
     @staticmethod
     def configure_input_size(
-        cfg, input_size_config: InputSizePreset = InputSizePreset.DEFAULT, model_ckpt: Optional[str] = None
+        cfg, input_size_config: InputSizePreset = InputSizePreset.DEFAULT, model_ckpt_path: Optional[str] = None
     ):
         """Change input size if necessary."""
-        input_size = get_configured_input_size(input_size_config, model_ckpt)
+        input_size = get_configured_input_size(input_size_config, model_ckpt_path)
         if input_size is None:
             return
 
@@ -211,79 +185,15 @@ class DetectionConfigurer(BaseConfigurer):
         logger.info("Input size is changed to {}".format(input_size))
 
 
-class IncrDetectionConfigurer(DetectionConfigurer):
+class IncrDetectionConfigurer(IncrConfigurerMixin, DetectionConfigurer):
     """Patch config to support incremental learning for object detection."""
 
-    def configure_task(self, cfg, train_dataset):
+    def configure_task(self, cfg, **kwargs):
         """Patch config to support incremental learning."""
-        super().configure_task(cfg, train_dataset)
-        if "task_adapt" in cfg and self.task_adapt_type == "mpa":
+        super(IncrConfigurerMixin, self).configure_task(cfg, **kwargs)
+        if "task_adapt" in cfg and self.task_adapt_type == "default_task_adapt":
             self.configure_task_adapt_hook(cfg)
 
-    def configure_task_adapt_hook(self, cfg):
-        """Add TaskAdaptHook for sampler."""
-        sampler_flag = True
-        if len(set(self.org_model_classes) & set(self.model_classes)) == 0 or set(self.org_model_classes) == set(
-            self.model_classes
-        ):
-            sampler_flag = False
-        update_or_add_custom_hook(
-            cfg,
-            ConfigDict(
-                type="TaskAdaptHook",
-                src_classes=self.org_model_classes,
-                dst_classes=self.model_classes,
-                model_type=cfg.model.type,
-                sampler_flag=sampler_flag,
-                efficient_mode=cfg["task_adapt"].get("efficient_mode", False),
-            ),
-        )
 
-
-class SemiSLDetectionConfigurer(DetectionConfigurer):
+class SemiSLDetectionConfigurer(SemiSLConfigurerMixin, DetectionConfigurer):
     """Patch config to support semi supervised learning for object detection."""
-
-    def configure_hook(self, cfg):
-        """Update cfg.custom_hooks."""
-        super().configure_hook(cfg)
-        # Set unlabeled data hook
-        if self.training:
-            if cfg.data.get("unlabeled", False) and cfg.data.unlabeled.get("otx_dataset", False):
-                if len(cfg.data.unlabeled.get("pipeline", [])) == 0:
-                    cfg.data.unlabeled.pipeline = cfg.data.train.pipeline.copy()
-                self.configure_unlabeled_dataloader(cfg)
-
-    @staticmethod
-    def configure_unlabeled_dataloader(cfg: Config):
-        """Patch for unlabled dataloader."""
-
-        model_task = {"classification": "mmcls", "detection": "mmdet", "segmentation": "mmseg"}
-        if "unlabeled" in cfg.data:
-            task_lib_module = importlib.import_module(f"{model_task[cfg.model_task]}.datasets")
-            dataset_builder = getattr(task_lib_module, "build_dataset")
-            dataloader_builder = getattr(task_lib_module, "build_dataloader")
-
-            dataset = build_dataset(cfg, "unlabeled", dataset_builder, consume=True)
-            unlabeled_dataloader = build_dataloader(
-                dataset,
-                cfg,
-                "unlabeled",
-                dataloader_builder,
-                distributed=cfg.distributed,
-                consume=True,
-            )
-
-            custom_hooks = cfg.get("custom_hooks", [])
-            updated = False
-            for custom_hook in custom_hooks:
-                if custom_hook["type"] == "ComposedDataLoadersHook":
-                    custom_hook["data_loaders"] = [*custom_hook["data_loaders"], unlabeled_dataloader]
-                    updated = True
-            if not updated:
-                custom_hooks.append(
-                    ConfigDict(
-                        type="ComposedDataLoadersHook",
-                        data_loaders=unlabeled_dataloader,
-                    )
-                )
-            cfg.custom_hooks = custom_hooks