+++ src/otx/algorithms/detection/configs/base/data/tiling/yolox_tile_pipeline.py
@@ -21,7 +21,6 @@ train_pipeline = [
         scaling_ratio_range=(0.1, 2),
         border=(-img_scale[0] // 2, -img_scale[1] // 2),
     ),
-    dict(type="MixUp", img_scale=img_scale, ratio_range=(0.8, 1.6), pad_val=114.0),
     dict(type="YOLOXHSVRandomAug"),
     dict(type="RandomFlip", flip_ratio=0.5),
     dict(type="Resize", img_scale=img_scale, keep_ratio=True),
@@ -46,15 +45,13 @@ train_pipeline = [
 ]
 
 test_pipeline = [
-    dict(type="LoadImageFromOTXDataset"),
     dict(
         type="MultiScaleFlipAug",
         img_scale=img_scale,
         flip=False,
         transforms=[
-            dict(type="Resize", keep_ratio=True),
+            dict(type="Resize", keep_ratio=False),
             dict(type="RandomFlip"),
-            dict(type="Pad", pad_to_square=True, pad_val=dict(img=(114.0, 114.0, 114.0))),
             dict(type="Normalize", **img_norm_cfg),
             dict(type="DefaultFormatBundle"),
             dict(type="Collect", keys=["img"]),