+++ tests/e2e/cli/semantic_segmentation/test_segmentation.py
@@ -86,7 +86,7 @@ else:
     templates_ids = [template.model_template_id for template in templates]
 
 
-class TestToolsMPASegmentation:
+class TestToolsOTXSegmentation:
     @e2e_pytest_component
     @pytest.mark.parametrize("template", templates, ids=templates_ids)
     def test_otx_train(self, template, tmp_dir_path):
@@ -280,7 +280,7 @@ args_semisl = {
 }
 
 
-class TestToolsMPASemiSLSegmentation:
+class TestToolsOTXSemiSLSegmentation:
     @e2e_pytest_component
     @pytest.mark.parametrize("template", templates, ids=templates_ids)
     def test_otx_train(self, template, tmp_dir_path):
@@ -307,12 +307,13 @@ class TestToolsMPASemiSLSegmentation:
 
 args_selfsl = {
     "--train-data-roots": "tests/assets/common_semantic_segmentation_dataset/train/images",
+    "--train-type": "Selfsupervised",
     "--input": "tests/assets/segmentation/custom/images/training",
     "train_params": ["params", "--learning_parameters.num_iters", "5", "--learning_parameters.batch_size", "4"],
 }
 
 
-class TestToolsMPASelfSLSegmentation:
+class TestToolsOTXSelfSLSegmentation:
     @e2e_pytest_component
     @pytest.mark.parametrize("template", templates, ids=templates_ids)
     def test_otx_train(self, template, tmp_dir_path):